+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditPanel.java
@@ -667,7 +667,6 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
     public void execute() throws ValidationException {
       if (confirmInsert()) {
         EntityEditModel.Insert insert = editModel().createInsert();
-        insert.validate();
         insert.notifyBeforeInsert();
         progressWorkerDialog(insert::insert)
                 .title(MESSAGES.getString("inserting"))
@@ -700,7 +699,6 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
     public void execute() throws ValidationException {
       if (confirmUpdate()) {
         EntityEditModel.Update update = editModel().createUpdate();
-        update.validate();
         update.notifyBeforeUpdate();
         progressWorkerDialog(update::update)
                 .title(MESSAGES.getString("updating"))